backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "supabase>=2.27.0",
    "upstash-redis>=1.5.0",
    "uvicorn>=0.40.0",
]
